修复安全扫描告警

+++ rosen/modules/texgine/src/font_descriptor_cache.cpp
@@ -524,11 +524,11 @@ int32_t FontDescriptorCache::WeightAlignment(int32_t weight)
         constexpr int kWeightDiffThreshold = Drawing::FontStyle::EXTRA_BLACK_WEIGHT / 2;
         if ((weight == Drawing::FontStyle::NORMAL_WEIGHT && item == Drawing::FontStyle::MEDIUM_WEIGHT) ||
             (weight == Drawing::FontStyle::MEDIUM_WEIGHT && item == Drawing::FontStyle::NORMAL_WEIGHT)) {
-            weightDiff = SPECIAL_WEIGHT_DIFF;
+            weightDiff = static_cast<uint32_t>(SPECIAL_WEIGHT_DIFF);
         } else if (weight <= Drawing::FontStyle::NORMAL_WEIGHT) {
-            weightDiff = (item <= weight) ? (weight - item) : (item - weight + kWeightDiffThreshold);
+            weightDiff = (item <= weight) ? static_cast<uint32_t>(weight - item) : static_cast<uint32_t>(item - weight + kWeightDiffThreshold);
         } else if (weight > Drawing::FontStyle::NORMAL_WEIGHT) {
-            weightDiff = (item >= weight) ? (item - weight) : (weight - item + kWeightDiffThreshold);
+            weightDiff = (item >= weight) ? static_cast<uint32_t>(item - weight) : static_cast<uint32_t>(weight - item + kWeightDiffThreshold);
         }
 
         // Retain the font weight with the smallest difference
